placed the roulette color & number generator to a separate helper file
